Fix suggested changes
