+++ .github/workflows/integration.yaml
@@ -19,5 +19,5 @@ jobs:
     steps:
       - uses: actions/checkout@v2
       - uses: actions/setup-go@v2
-      - run: export MAINNET=true && make test-integration
+      - run: MAINNET=true make test-integration
